ADD vite alias to images dir, icons lib
We can use '@img' to refer to the './resources/img' directory for when
we need to use image assets. Example use is in the header logo.

Can now use icons as <i-(iconSet)-(iconName)>, e.g.:
<i-mdi-view-dashboard-variant /> is used as an example in the nav bar
Dashboard link.

Unplugin-icons is used as it allows us to forgo Iconify's default
behaviour of retrieving icons over the internet on use. It lets us store
the icons locally and will also only bundle used icons in prod builds.
Unplugin-vue-components is for auto importing the icon on use, otherwise
we'd have to manually import each icon as a component before using it.

Instead of downloading every supported icon set in Iconify, I've only
downloaded Material Design Icons (mdi) and OpenMoji (openmoji) to save
space & bandwidth during CI/CD builds.

Not sure we want to stick with mdi icons as the general icons, just
using it cause it seems pretty standard.