ADD pinia lib for vue state storage
Pinia is how Vue3 shares state among components, it's the successor to
the Vuex library from Vue2.

As an example implementation, I added an api call that gets the iPeer
version. There's a corresponding store that retrieves this via Axios and
the info is used in the App component.

I decided to keep using Axios instead of the new Fetch API, as Axios
seems to automatically handle a lot of header stuff for you, and Fetch
requires more manual configuration.

The Axios plugin should give us a singleton Axios instance that can be
used across the entire UI. While components can access Axios via the
root app, the Pinia stores need to import the Axios plugin.