ADD quasar framework for frontend ui
Quasar is a Vue based UI framework that allows one codebase to create
web app, desktop app, or mobile app all at once. However, the way I've
integrated it into Laravel means we're not using the Quasar CLI, so it's
web only unless we decide to split off the UI away from Laravel.

The hope is that because Quasar has a lot of ready made UI components,
we can quickly get this working, even if it it doesn't look like we
want. I do miss the versaility of Tailwind's utility classes though.

I've set iPeer purple as the primary color and tried to pick a green
that might complement the purple as secondary.

It seems that because this was made for full screen apps, we don't have
anything facility for a fixed width layout like we had previously. So I
gave up on that.