ADD laravel sanctum for auth, basic login page
Laravel Sanctum is an SPA authentication library. It can issue API
tokens or use Laravel's built in cookie auth. Currently, we're using the
cookie auth, but this requires us to enable sessions on the API. Not
sure if this is the best approach or if we should use tokens only so
it's stateless.

Basic register page to create an account. This is only for dev purposes,
we should restrict or disable it later.

Basic login page, just enter username/password and checks if the login
request is successful.

No user roles system yet, so everyone has access to everything.

There's a /login API call, but since there's not auth required
endpoints, it's kinda pointless right now. There's also a basic test
case for the endpoint.