ADD allow string true/false to validate as boolean
The default Laravel boolean validation treats 'true' and 'false' strings
as invalid input. This was inconvenient since javascript converted true
and false to string 'true' and 'false'. So I've created a custom
validation rule BoolStr for such cases.

BoolStr is only a validation rule though, it doesn't actually change the
input. This means that 'true' and 'false' string will remain string. So
there's also an accompanying 'toBoolean' function that actually converts
it to a boolean.

This means that after passing validation, you MUST pass it through the
toBoolean() call to actually get a bool.