EDIT ui all pages should use q-page as root element
This seems like an easy way to classify whether a component should be in
pages/ or in components/. The q-page component needs to be inside a
q-page-container, this is done in App.vue, where the <router-view />
should be replaced by whater q-page is on the current route.

Strangely, it looks like q-footer doesn't need to be ahead of
q-page-container. I encountered some z-height issues with features that
use absolute positioning when q-page-container wasn't near the end of
the body, so I've moved it. Didn't end up using the absolute position
(always on top) feature though.