ADD ui error store and error box for error display
This is another attempt at centralizing a common function, where we
hopefully won't have to duplicate error handling in every component.

The ErrorStore is meant to process Laravel errors and present it in a
human readable form.

The ErrorBox component is meant to be placed in all forms and it'll
display the errors encountered in a list.