EDIT ErrorStore handles field validation, 404 errs
Laravel returns 422 for field validation errors. ErrorStore will process
it into a form that'll show up in ErrorBox and also can be used by form
inputs to display errors. An example is given in `DEVELOPMENT.md`.

ErrorStore will also handle 404 errors, adopting the error message from
the server if available, a generic message otherwise.

ErrorStore handling 409 errors should be rare, like if we try to delete
a user, but the user was already being deleted.

A general catchall was added, but hopefully we won't have to use it.

The UserController API was changed so it'll properly throw 404 errors
when someone calls show() and destroy() on non-existent users. Also
throws 409 error if, for some reason, destroy() failed to delete a user.
From the Laravel code, it seems like this should only happen if user is
already gone or if user was already being deleted.