EDIT ui login, loading status, no success notify
Removed the 'Login Successful' toast, it's just unnecessary since it
should be pretty obvious that login was successful.

Added a loading status to the login form so users know it's processing.

Didn't like the look of the customized secondary color, so disabled the
customized css to go back to stock.