FIX ui ErrorStore can't handle no server resp errs
The ErrorStore was unable to handle errors where the server doesn't
return a response. So I added handling for errors where we sent a
request to the server but received no response and where no request was
sent. Don't have specific error in mind, so the error handling for these
two new cases are pretty generic.

I accidentally came across the last one, where requests aren't sent,
with a ReferenceError where a variable name was used without being
declared. So built it around that case. Would've liked to be able to
display the stack trace more nicely, but the naive solution of creating
a html text node doesn't work. So I think it has to be slightly more
complicated, perhaps requires adding another element to display it in
ErrorBox.

Refactored out the http response code handling to it's own function,
since it was pretty long and made the logic harder to follow.