ADD ui UserStore to persist UserList data
When jumping from UserList to a UserInfo page and then clicking Back to
UserList, you lose the page that you were on in UserList. To persist the
state of UserList, I've moved the pagination and users storage to the
UserStore.

Decided to try out the async/await syntax for axios calls. Apparently,
it's just syntactic sugar for promises, but it does seem more clean than
the promise then/catch syntax. Hopefully remembering to declare async
for functions that call axios won't be too much of an issue.

There were suggestions that business logic should live in a separate
service layer from stores. But I'm not sure we need that kind of
separation yet, so I've left the business logic inside the store.

Not sure how it works, but v-model can apparently be now set to a
store's state directly? It seems to work fine for now, but I'm not clear
on the underlying mechanism, so maybe it'll come back to bite us later.