ADD ui UserList search filter
This is performing a WHERE LIKE query on username, name, and email
fields. I kinda worry that the performance on this query might be too
slow as the database grow in the future.

Added a escapeLike() function that'll escape the search term for the
LIKE query for % and _ chars that has special meaning just for LIKE
queries. This was added in the same way as the toBoolean() function and
is available globally. I found whitespace issues in the toBoolean file
and reformatted it.

I considered adding the filter variable to the UserStore's pagination,
but decided to make it separate as we took the pagination structure from
the q-table component, and filter is a separate variable according to
q-table. It should be less confusing to follow q-table's lead.

Also added a test case for the filter param.