EDIT error system messages moved to components
Big overhaul of the error system. I felt that having to manually clear
error messages while handling axios calls was a bit clunky, so wanted a
system that just hides old errors while displaying only the newest
errors.

ErrorStore now splits errors into two categories: latest and old. Latest
errors are most recent errors that are displayed by default to the user.
We treat all errors that happen within 1 second of the first error as in
the same 'error frame'. When errors come in that are newer than the
current error frame, the existing latest error array is moved to old
errors. Old errors are in descending order, with the first element being
the newer error frame, e.g.: [[error5, error4],[error3],[error2, error1]]

When shown by the ErrorBox, old errors are collapsed by default, but can
be accessed by the user if necessary. There's a clear errors button available to the user.

Errors are still automatically cleared on page change.

During this overhaul, I found that rather than the errorStore generating
the error messages, it's better to farm that out into dedicated error
message components. This gives us more flexibility, such as displaying
stacktraces in a preformatted element, that we couldn't have before.

The ErrorCard component is the generic error message card underlying the
rest of the error cards.

ServerErrorCard handles HTTP 400/500 errors.

NoResponseErrorCard handles when server didn't respond. This is
currently just a hardcoded error message.

CodeErrorCard is all other errors and assumed to be something wrong with
the code (syntax, variable names, etc). This will output an ugly
stacktrack of no use to regular users, but hopefully lets them copy it
as bug reports.

Renamed global axios to $axios to conform to $ naming of custom globals.