EDIT ui UserInfo delete and edit user btn working
Moved getting user info and deleting user into the UserStore, so the
UserStore will issue axios requests. We can let exceptions bubble up for
the generic error catcher to put out the right error messages into
ErrorBox.

The Edit user button just sends the user to the UserForm, we'll still
need to make UserForm understand Edit actions.

Added Quasar's breakpoint aware flex/spacing classes, as I wanted the
Delete button to be less accessible. On mobile, the Delete button needs
to be on the left to be less accessible, while it needs to be on the
right for desktop. This needed the breakpoint aware classes to
implement.