EDIT ui UserForm now supports edit operations
If a user ID is present in the route, then UserForm will assume we're
editing an existing user.

Abstracted out the password field into its own PasswordInput component.
I was going to use this to implement a verify new password field for
when changing passwords, but it's not really necessary for an early
prototype yet. This also replaced the password field in LoginPage.

Added support for 403 Forbidden errors.