ADD ui course store, components, page
Can now list/create/edit/delete courses.

For the CourseStore, decided that we don't need the 'Course' part of
'getCourse', so the actions are now renamed to get(), create(), edit(),
remove(). Pretty sure I could've used new() instead of create() and
delete() instead of remove(), but new and delete being javascript
keywords seems like it might be safer to avoid them.

It turns out the Store and List/Form/Info components duplicate a lot of
code with their User counterpart. So this is another opportunity for
refactoring later.