ADD custom req PaginatedRequest for paging params
Trying to generalize pagination requests so that we don't have so much
code duplication where every model controller needs its own set of
pagination code. AbstractPaginatedRequest should be used as the parent
class for model specific requests. Naming the class Abstract is meant to
signal that the class shouldn't be used directly, only as a parent
class.

E.g.: UserPaginationRequest extends PaginatedRequest and is the
pagination request for the UserController.

The pagination params that needs to be present should be the same for
all pagination requests. The difference is that the fields that can be
sorted by will vary from model to model. So model specific
PaginatedRequest will override the protected array $sortFields for the
fields that are relevant for that model.

The controller using these request objects just need to grab the
validated data from the request. The router will have executed the
validation before it even gets to the controller.