ADD logout function to controller and ui banner
I was deleting the session from the database in order to test logouts.
Got tired of it so decided to put in a basic logout button in the UI.

Logout API request goes to /logout and is expected to be an empty POST
request. It's a POST request cause apparently some page caching
mechanisms might hit the URL with a GET and cause an unintended logout.