ADD enrol users into courses via CourseUser table
The CourseUser table is treated as a pivot table by the ORM. There's
facilities to use it without adding an explicit model. So I'm trying to
use the built-in facilities that modifies the table indirectly. This is
mainly the attach/detach/sync methods. attach/detach is just add and
remove respectively.

Sync is slightly more complicated in that it'll do both remove and add
to make CourseUser match the given state.

Couldn't think of a current use case for CourseUser show/update. So I've
left them returning 501 unimplemented for now. Show might be useful to
check if a user is enroled in a course, update might be useful for
changing a user's role. But I'll wait until actually needing it before
implementing.

Tried to make a generic paginated index test, but found that I don't
have a good enough grasp of what can be abstracted yet.