ADD ui can now enrol/drop user from courses
UserSearchStore is meant for searching for a user to enrol into a
course. I first tried modifiying the existing UserList and UserStore to
support both use cases. But it got complicated enough that it seems
creating another store and component was the easier solution for now.

Both CourseUserList and CourseUserForm can only enrol/drop one user at a
time. This is because while the api for enrol supports multiple users,
the drop api only does this one at a time. So for simplification, I only
implemented one user at a time for both.

Not sure this is the best organization, since we might also want to be
able to go a UserInfo and edit their course list that way. But we'll see
how it works out for now.

I was thinking that CourseUserForm should require the user to search for
a user before showing any users. This means that you can't browse the
entire user list looking for the user you want. This is pretty much what
you have to do once you get a large amount of users. Not sure how
this'll work out either.

Added a resetPagination() to the AbstractPaginatedStore. This is meant
to clear all pagination state back to defaults.