ADD PermissionSeeder to seed default roles & perms
We have two broad types of roles & permissions: System or course.
System permissions apply to the entire app. Course permissions apply
only inside the course you're enroled in. For example, an admin is a
system role and has access to every. A student is a course role and they
can only see courses they're enroled in. In older iPeer 3, there is no
such separation, and there are only system roles. So we had issues where
some users (TAs) were students in some courses but instructors in others
needed to have two separate accounts for each of those roles. The
course-specific roles are meant to address this.

Three default roles:
admin - system role, access everything
instructor - course role, only access courses they're in
student - course role, only access courses they're in

I'm thinking of using the wildcard permission feature of
laravel-permission to check for course-specific permissions. This means
that we'll need role and permission "templates" to copy from, changing
the copy so that it limits to a specific courseId.

Also added some custom columns to the stock laravel-permission
migration.

* Role "is_system", indicates system role
* Role "is_template", template role

Added a role_id column to the CourseUser table, which will store the
user's course role. I'm trying to see if we can get away with only
storing the course role in CourseUser and not in laravel-permission
tables. It would make laravel-permission easier to use if we also let it
store course role assignments, but I'm wondering if there might be
issues keeping the CourseRole and laravel-permission roles synced, so
experimenting to find out.

Needed to use our own custom Role and Permission models since I wanted
to add some helper functions to them.