ADD user/course policies to check permission, test
Important: For automatically applying policy to a controller, the
controller's method signatures need to match what's expected. I
initially had nothing working because instead of taking in model
instances like User or Course, I was taking in an int id. The automatic
policy only started working when I switched to taking the model instance
like the policies expect.

Added a Gate to let admins have access to everything.

PermissionTest checks that the policies are actually working.