FIX test case requests fail with unauthorized
Added AbstractApiTestCase as the parent class of API tests. This
basically just runs the the database seeder so that permissions and
roles are seeded.

Added admin user role to the users used in tests so that they should
have no issue with any endpoints.

Moved PermissionTest to Api/ since it tests API endpoints.