ADD CourseUser pivot model
With the addition of role_id to the CourseUser table, the course_user
pivot table manages more than 2 relations. I don't see an easy way to
manage this with the built-in pivot functions so I want to stop relying
on only the built-in pivot model we were previously using for
CourseUser.

Queries for which we need to reference all 3 relations in CourseUser
(course, user, role) seems to be easier using the CourseUser model
direclty rather than indirectly refering to the pivot table via
course/user's relations.