EDIT CourseUserController return role & auth ctrl
CourseUserController will now also return the role. Since this is now
using the CourseUser model, the id is now the CourseUser table id.
Previously, it was the user id.

Also added access control. So only instructors and above should be able
to get and change enrolment.

Updated the CourseUserControllerTest to test that students aren't
allowed to do operations and to accomodate roles.

CourseUserPolicy doesn't follow Laravel convention in the params we're
expecting for each policy call, so we're using "$this->authorize()" in
the controller methods to explicitly call the policy with the param we
need.