ADD CourseRoleController to return course roles
Course roles such as 'student' and 'instructor' for assignment to new
users in a course.

Due to how the permission system works, each course has their own copy
of the student and instructor role. This is why it's a CourseRole
controller.

Since the role name isn't user friendly, I added a "display_name" column
to the Role table for a friendly name to show to users.

Added some helpers to more easily get course roles.

Since we don't expect there to be pages and pages of course roles, we re
not using a paginated request for the index(). We're just returning all
roles in a single request. No other methods are implemented, but it's
leaving room for custom course roles later one if we want to do that.