ADD select course role when adding user to course
You can now choose a course role when adding a new user to a course.

There is now a CourseRoleSelect component that lists the course roles
available.

I noticed that Quasar has a Field wrapper that lets you use v-model to
pass on the original input's value to the component above it. This makes
it easier to customize Quasar input components. It turns out fields like
QInput and QSelect are already wrapped this way. So my custom password
field didn't need to use the model-value prop to pass value back up. As
the CourseRoleSelect uses a QSelect, I've used the v-model for that one,
but kept the password input using the old method for reference.