MOVE PaginatedIndex and CanCourse into traits
Removed AbstractApiController as it was only used to add the
paginatedIndex() method. Realized that this method is a much better fit
as a trait. For some reason, I errorneously thought that traits was some
Laravel backend thing. This is under the Controller\Traits namespace as
I think only controllers will use this.

Removed ApiPolicy as it was only used to add the canCourse() method.
Also moved this into its own trait. This is under the App\Traits
namespace as both Policies and Controllers will use it.

Maybe both should just be automatically used in Controller.php later on
down the road, but I'm not sure if that's a good refactor yet.