ADD CreateSecondCourseWithUsers trait for tests
Trying out moving creating test data into traits. When we need a second
course with users, we can use the CreateSecondCourseWithUsers trait and
access the data created with:

    $this->course2
    $this->course2InstructorRole
    $this->course2StudentRole
    $this->course2Instructor
    $this->course2Student

This uses the same mechanism as Laravel's test traits like
RefreshDatabase, where the trait's setup method is called in the
setUpTraits() method. Since setUpTraits() is called before setUp(), I've
moved all the stuff that used to be in setUp() into setUpTraits(), as we
need the database to be seeded with permissions in order for our custom
traits to work.