ADD DateTimeIso8601 rule EDIT BoolStr deprecation
To avoid repeating all the acceptable ISO8601 formats, I've copied the
DateTimeIso8601 validation rule from the Laravel JSON:API. Might be a
good idea to adopt the library in its entirety, but it seems to be a
complex enough change that I'm not sure we need it... yet.

The validation Rule parent class is apparently deprecated, so I've moved
our rules to use it.

BoolStr is really only there for the custom error message and the new
static method for checking if a value can be safely interpreted as a
boolean. I'm creating a PrepareBool trait that's meant to be used in
prepareForValidation() that'll safely convert the boolean before it gets
to BoolStr.