ADD PrepareBool trait MOVE PaginatedIndex trait
PaginatedIndex was in the app/Http/Controllers/Traits/ directory. I've
moved it so it's with all the other app traits in app/Traits/. Only
other traits directory is the traits used for tests and those shouldn't
be used in the app.

The PrepareBool trait converts values that can be safely intepreted as
booleans into actual booleans. E.g.: string 'true' or integer '1' is
converted to boolean true. It'll avoid modifying invalid or null values.
The intent is to use prepareBool() in prepareForValidation() on all
boolean params. Invalid or null values are passed on to the validation
rules to either throw an error or accept.