ADD Assignment table & controller & tests
The create_assignments_tables migration also created the
AssignmentQuestions table which will be implemented later. I'm thinking
that later on we'll have a separate Evaluation table that is almost
identical to the Assignment table where we store assignment templates.
This is to match features with iPeer3. When an assignment is create, we
copy all the params from the evaluation template.

Updated PermissionSeeder for Assignment permissions. Assignment
permissions are course permissions, there is only 3 currently:
- viewPublishedAssignment: can view published assignments
- viewUnpublishedAssignment: can view unpublished & published
  assignments
- editAssignment: can create, edit, delete assignments

Note that while we say viewUnpublishedAssignment can view both published
and unpublished assignments, I initially implemented
AssignmentPolicy->view() in such a way that it only checks for
viewPublishedAssignment permissions and so failed to allow instructors
to view published assignments. AssignmentPolicy->view() no longer
behaves this way, but So I've given viewPublishedAssignment permission
to instructors as well to prevent this type of issue.

The editAssignment permission is standing in for other permissions for
now until we see a need to separate them out.

AssignmentController's store and update have custom requests for
permission checking and for param validation: StoreAssginmentRequest and
UpdateAssignmentRequest. The skeleton was autogenerated by artisan.

I decided not to worry about AssignmentController's delete returning
false. If model delete returns false, it means a delete is already in
progress. Either way, the model is being deleted, so we shouldn't return
an error like in UserController. The scenario is difficult to replicate
in tests anyway, so we can't be sure it'll work.

AssignmentControllerTest is using the CreateAssignments trait to create
test data. CreateAssignments create a published and a unpublished
assignment.